backend/requirements_deploy.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.2
bcrypt==4.1.3
mysql-connector-python==8.4.0
gunicorn==21.2.0
python-dotenv==1.0.1
markupsafe>=2.1.5
